User: Move auth views into new module
